another baby step
I'm still working through in my brain where I want to store what data. I need this to be efficient, where it is easy to get what vertices are adjacent. or how to get from vertex A to B.
I'm just throwing in some functions to the classes as I explore what data needs to be stored where.