more implementation of Graph's functions
The main goal with this commit (and the previous commits) was to make sure that a user of this library only has the ability to interface with the structure of the graph through their instance of the Graph.
That way, we can make sure that vertices and edges are only added in the correct way.

This will prevent edges and vertices from existing that don't have the appropriate object references or have incorrect object references.